Change variable type at Geo class from float to Double for accepting null

+++ src/main/java/com/example/feature/level1/domain/Geo.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @Setter
 public class Geo {
-    private float lat;
-    private float lng;
+    private Double lat;
+    private Double lng;
 }
